Verificaciones Modulos Requisiciones Sistemas

+++ database/migrations/2022_03_04_100238_create_requisiciones_sistemas_table.php
@@ -18,7 +18,7 @@ class CreateRequisicionesSistemasTable extends Migration
             $table->unsignedBigInteger('IdUser'); //Id de Session
             $table->date('Fecha');
             $table->integer('Folio');
-            $table->integer('Estatus')->default(0);
+            $table->tinyInteger('Estatus')->default(0);
 
             $table->unsignedBigInteger('Perfil_id')->nullable();
             $table->foreign("Perfil_id")->references("id")->on("perfiles_usuarios")
